test: clean mock data
- refine auto receive service
- clean mock ledger code
