+++ rpc/api/util_test.go
@@ -39,7 +39,7 @@ func setupUtilAPI(t *testing.T) (func(t *testing.T), *process.LedgerVerifier, *U
 	api := NewUtilAPI(l)
 
 	var blocks []*types.StateBlock
-	if err := json.Unmarshal([]byte(MockBlocks), &blocks); err != nil {
+	if err := json.Unmarshal([]byte(mock.MockBlocks), &blocks); err != nil {
 		t.Fatal(err)
 	}
