+++ vm/contract/abi/mock_ledger_test.go
@@ -8,12 +8,13 @@
 package abi
 
 import (
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
+
+	"github.com/qlcchain/go-qlc/mock"
 
 	"github.com/google/uuid"
 
@@ -24,12 +25,7 @@ import (
 )
 
 var (
-	// qlc_3pekn1xq8boq1ihpj8q96wnktxiu8cfbe5syaety3bywyd45rkyhmj8b93kq
-	priv1, _ = hex.DecodeString("7098c089e66bd66476e3b88df8699bcd4dacdd5e1e5b41b3c598a8a36d851184d992a03b7326b7041f689ae727292d761b329a960f3e4335e0a7dcf2c43c4bcf")
-	// qlc_3pbbee5imrf3aik35ay44phaugkqad5a8qkngot6by7h8pzjrwwmxwket4te
-	priv2, _ = hex.DecodeString("31ee4e16826569dc631b969e71bd4c46d5c0df0daeca6933f46586f36f49537cd929630709e1a1442411a3c2159e8dba5742c6835e54757444f8af35bf1c7393")
-	account1 = types.NewAccount(priv1)
-	account2 = types.NewAccount(priv2)
+	account1 = mock.Account1
 )
 
 func setupLedgerForTestCase(t *testing.T) (func(t *testing.T), *ledger.Ledger) {
@@ -42,31 +38,9 @@ func setupLedgerForTestCase(t *testing.T) (func(t *testing.T), *ledger.Ledger) {
 		t.Fatal(err)
 	}
 	l := ledger.NewLedger(cm.ConfigFile)
-	//ctx := vmstore.NewVMContext(l)
-	//verifier := process.NewLedgerVerifier(l)
-	//
-	//for _, v := range common.genesisInfos {
-	//	mb := v.GenesisMintageBlock
-	//	gb := v.GenesisBlock
-	//	err := ctx.SetStorage(types.MintageAddress[:], v.GenesisBlock.Token[:], v.GenesisBlock.Data)
-	//	if err != nil {
-	//		t.Fatal(err)
-	//	}
-	//	if b, err := l.HasStateBlock(mb.GetHash()); !b && err == nil {
-	//		if err := l.AddStateBlock(&mb); err != nil {
-	//			t.Fatal(err)
-	//		}
-	//	}
-	//	if b, err := l.HasStateBlock(gb.GetHash()); !b && err == nil {
-	//		if err := verifier.BlockProcess(&gb); err != nil {
-	//			t.Fatal(err)
-	//		}
-	//	}
-	//}
-	//_ = ctx.SaveStorage()
 
 	var blocks []*types.StateBlock
-	if err := json.Unmarshal([]byte(MockBlocks), &blocks); err != nil {
+	if err := json.Unmarshal([]byte(mock.MockBlocks), &blocks); err != nil {
 		t.Fatal(err)
 	}
